refactor(rest): share power provisioning validation
